Update calc methods to reflect property changes
